tapcli: move `proof import` command into new dev subcommand
