+++ cmd/tapcli/main.go
@@ -336,6 +336,7 @@ func main() {
 	app.Commands = append(app.Commands, addrCommands...)
 	app.Commands = append(app.Commands, proofCommands...)
 	app.Commands = append(app.Commands, universeCommands...)
+	app.Commands = append(app.Commands, devCommands...)
 
 	if err := app.Run(os.Args); err != nil {
 		fatal(err)
