+++ cmd/tapcli/tapdevrpc_default.go
@@ -1,0 +1,7 @@
+//go:build !dev
+
+package main
+
+import "github.com/urfave/cli"
+
+var devCommands []cli.Command = nil
